+++ src/main/java/com/w3gdata/actionblock/ChangeSelection.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class ChangeSelection implements ActionBlock {
-    public static final int ID = 0x16;
+    public static final byte ID = 0x16;
 
     public static final int ADD_TO_SELECTION = 0x01;
