+++ src/main/java/com/w3gdata/actionblock/GameSpeedIncreasing.java
@@ -1,5 +1,5 @@
 package com.w3gdata.actionblock;
 
 public class GameSpeedIncreasing implements ActionBlock {
-    public static final int ID = 0x04;
+    public static final byte ID = 0x04;
 }
