+++ src/main/java/com/w3gdata/actionblock/ItemGivenDropped.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class ItemGivenDropped extends UnitBuildingAbilityTargeted {
-    public static final int ID = 0x13;
+    public static final byte ID = 0x13;
 
     public int targetObjId1;
