+++ src/main/java/com/w3gdata/actionblock/PauseGame.java
@@ -1,5 +1,5 @@
 package com.w3gdata.actionblock;
 
 public class PauseGame implements ActionBlock {
-    public static final int ID = 0x01;
+    public static final byte ID = 0x01;
 }
