+++ src/main/java/com/w3gdata/actionblock/ResumeGame.java
@@ -1,5 +1,5 @@
 package com.w3gdata.actionblock;
 
 public class ResumeGame implements ActionBlock {
-    public static final int ID = 0x02;
+    public static final byte ID = 0x02;
 }
