+++ src/main/java/com/w3gdata/actionblock/SaveGame.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class SaveGame implements ActionBlock {
-    public static final int ID = 0x06;
+    public static final byte ID = 0x06;
     public String saveGameName;//null terminated string
 
     public SaveGame(String saveGameName) {
