+++ src/main/java/com/w3gdata/actionblock/SelectGroupHotkey.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class SelectGroupHotkey implements ActionBlock {
-    public static final int ID = 0x178;
+    public static final byte ID = 0x178;
 
     public int groupNumber;
