+++ src/main/java/com/w3gdata/actionblock/SelectSubgroup114b.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class SelectSubgroup114b implements ActionBlock {
-    public static final int ID = 0x19;
+    public static final byte ID = 0x19;
 
     public int itemId;
