+++ src/main/java/com/w3gdata/actionblock/TransferResources.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class TransferResources implements ActionBlock {
-    public static final int ID = 0x51;
+    public static final byte ID = 0x51;
 
     public int playerSlotNumber;
