+++ src/main/java/com/w3gdata/actionblock/UnitBuildingAbility2Targets2Items.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class UnitBuildingAbility2Targets2Items extends UnitBuildingAbilityTargeted {
-    public static final int ID = 0x14;
+    public static final byte ID = 0x14;
 
     public static final int UNKNOWN_BYTES_SIZE = 9;
