+++ src/main/java/com/w3gdata/actionblock/UnitBuildingAbilityTargeted.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class UnitBuildingAbilityTargeted extends UnitBuildingAbility {
-    public static final int ID = 0x11;
+    public static final byte ID = 0x11;
 
     public int targetX;
