+++ src/main/java/com/w3gdata/actionblock/Unknown0x1B.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class Unknown0x1B implements Unknown {
-    public static final int ID = 0x1B;
+    public static final byte ID = 0x1B;
 
     public static final int ADDITIONAL_BYTE = 0x1A;//for WarCraft III patch version <= 1.14b
