+++ src/main/java/com/w3gdata/actionblock/Unknown0x6B.java
@@ -3,7 +3,7 @@ package com.w3gdata.actionblock;
 import com.w3gdata.util.ByteBuffer;
 
 public class Unknown0x6B implements Unknown {
-    public static final int ID = 0x6B;
+    public static final byte ID = 0x6B;
 
     @Override
     public int getSize() {
