+++ src/main/java/com/w3gdata/actionblock/Unknown0x6C.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class Unknown0x6C implements Unknown {
-    public static final int ID = 0x6C;
+    public static final byte ID = 0x6C;
 
     @Override
     public int getSize() {
