+++ src/main/java/com/w3gdata/actionblock/Unknown0x74.java
@@ -1,7 +1,7 @@
 package com.w3gdata.actionblock;
 
 public class Unknown0x74 implements Unknown {
-    public static final int ID = 0x74;
+    public static final byte ID = 0x74;
 
 
     @Override
